:recycle: [refactor] : contents/discovery dto 수정 및 반영
