+++ components/discovery/Discovery.tsx
@@ -4,15 +4,6 @@ import Link from "next/link";
 import React, { HTMLAttributes } from "react";
 import CommonPillButton from "../common/CommonPillButton";
 import EmptyImage from "../common/EmptyImage";
-import DiscoveryItem from "./DiscoveryItem";
-
-enum ContentsType {
-  "Youtube" = "유튜브",
-  "Homepage" = "홈페이지",
-  "Blog" = "블로그",
-  "Link" = "링크",
-  "Profile" = "프로필",
-}
 
 type Props = { className?: string } & HTMLAttributes<HTMLDivElement>;
 const Discovery = async ({ className, ...props }: Readonly<Props>) => {
@@ -51,14 +42,14 @@ const Discovery = async ({ className, ...props }: Readonly<Props>) => {
           {dailyDiscoveryData.map(
             (item, index) =>
               index > 0 &&
-              index < 3 && (
+              index < 4 && (
                 <React.Fragment key={item.id}>
-                  <DiscoveryItem
-                    contentType={ContentsType[item.urls[0].type] || "기타"}
-                    title={item.title || ""}
+                  {/* <DiscoveryItem
+                    contentType={ContentsType[item.urls[0].url] || "기타"}
+                    title={item. || ""}
                     imageUrl={item.image}
-                  />
-                  {index < 3 - 1 && <div className="border-b" />}
+                  /> */}
+                  {index < 3 && <div className="border-b" />}
                 </React.Fragment>
               )
           )}
